backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-jwt-extended
flask-cors
psycopg2-binary
pydantic
redis
celery
apscheduler
gunicorn
python-dotenv
